% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_36_10era.R
\name{f_36_10era}
\alias{f_36_10era}
\title{36an oinarritutako zenbakikuntzatik 10erakora}
\usage{
f_36_10era(x)
}
\arguments{
\item{x}{Karaktere motakoa. 36an oinarritutako zenbakia (zenbaki eta letraz osatua izan liteke)}
}
\description{
Funtzinoiak 36an oinarritutako  \code{x} zenbakia sistema hamartarrera dakar.
}
\details{
Funtzino hau behar da \code{f_idat} funtzino barruan.
}
\examples{
f_36_10era('11')
f_36_10era('ab2')
}
