% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ekitaldiak.R
\name{f_ekitaldiak}
\alias{f_ekitaldiak}
\title{Idazteko orduan egin diren ekitaldiak identifikatzen ditu}
\usage{
f_ekitaldiak(objektua, muga = 1000)
}
\arguments{
\item{objektua}{IDAT erabilita sortutako zerrenda edo data framea}

\item{muga}{Ekitaldi bakoitza zehazteko kontuan izateko gehienezko muga, milisegunduetan}
}
\value{
Etherpaden idatzitako dokumentu baten \code{objektua}aren \code{muga} baten araberako ekitaldiak
}
\description{
Idazteko orduan egin diren ekitaldiak identifikatzen ditu
}
\examples{
require(dplyr)
# o_idat <- f_IDAT(fitxategia, kokapena)
f_ekitaldiak(o_idat)$df_aldaketak |> select(ekitaldia)

}
