% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_epoc_ekarri.R
\name{f_epoc_ekarri}
\alias{f_epoc_ekarri}
\title{EPOC formatutik POSIXera ekarri}
\usage{
f_epoc_ekarri(x)
}
\arguments{
\item{x}{Zenbakizkoa motakoa. 1970-01-01tik ona igarotako denbora, milisegunduetan eta}
}
\description{
Funtzinoiak zerbitzari aldeko EPOC formatuko datua \code{x}
eredu irakurgarrira dakar
}
\details{
Funtzino hau behar da \code{f_idat} funtzino barruan.
}
\examples{
f_epoc_ekarri(1600352447019)
}
