% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_garbitu.R
\name{f_garbitu}
\alias{f_garbitu}
\title{Analisietan oztopo egin ohi duen informazio lerroak ezabatzeko zenbait aukera}
\usage{
f_garbitu(
  objektua,
  copypaste = FALSE,
  n = 7,
  hutsa = TRUE,
  ni = TRUE,
  nirea = "a.WqkJLdiHlrEQ40Fl"
)
}
\arguments{
\item{objektua}{IDAT erabilita sortutako zerrenda edo data framea}

\item{copypaste}{Luzera batetik gorako `revs` ezabatzen ditu. Luzera `n`k zehazten du}

\item{n}{Ezabatuko den informazioaren gutxieneko karaktere kopurua. Balio lehenetsia `7` da}

\item{hutsa}{Identifikatu gabeko `author`ik agertzen ez diren lerroak ezabatzen ditu `TRUE`k.
Hori zerbitzarian aurretik egon ohi den informazioa izaten da.}

\item{ni}{Paketearen egileak erabili ohi duen "author" ezabatzen du}

\item{nirea}{Analisia egiten ari denaren edo beste baten erabiltzaileak idatzitakoa
ezabatzen du. Balio lehenetsia paketearen egilearen `author` balioa da}
}
\value{
Etherpaden idatzitako dokumentu baten \code{objektua}ari kentzen dizkio
informazio nahasiko lerroak, \code{copypaste} egindakoa delako, zein `author` identifikazio
\code{hutsa} edo \code{ni} nazelako
}
\description{
Analisietan oztopo egin ohi duen informazio lerroak ezabatzeko zenbait aukera
}
\examples{
require(dplyr)
# o_idat <- f_IDAT(kokapena, fitxategia)

f_garbitu(o_idat)$df_aldaketak

}
