% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_IDAT.R
\name{f_IDAT}
\alias{f_IDAT}
\title{Etherpad dokumentu batean idatzitakoa aztertzeko oinarrizko data framea sortzen du}
\usage{
f_IDAT(kokapena = kokapena, fitxategia = fitxategia)
}
\arguments{
\item{kokapena}{Etherpad fitxategia non dagoen; e.b. "/home/juan/dokumentuak/"}

\item{fitxategia}{Etherpad fitxategiaren izen osoa, luzapena eta guzti. "adibidea.etherpad"}
}
\value{
Etherpaden idatzitako dokumentuaren analisirako, \code{kokapena}ren eta \code{izena} emanda.
}
\description{
Etherpad dokumentu batean idatzitakoa aztertzeko oinarrizko data framea sortzen du
}
\examples{
# kokapena <- 'extdata/'
# fitxategia <- 'txepetx-2-lo1uvq90o.etherpad'
# o_idat <- f_IDAT(kokapena, fitxategia)
o_idat

}
