% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_IDAT2.R
\name{f_IDAT2}
\alias{f_IDAT2}
\title{Etherpad dokumentu batean idatzitakoa aztertzeko oinarrizko data framea sortzen du
Shiny app batean erabiltzeko egokitua}
\usage{
f_IDAT2(fitxategia = fitxategia)
}
\arguments{
\item{fitxategia}{Etherpad fitxategiaren izen osoa, luzapena eta guzti. "adibidea.etherpad"}
}
\value{
Etherpaden idatzitako dokumentuaren analisirako, \code{kokapena}ren eta \code{izena} emanda.
}
\description{
Etherpad dokumentu batean idatzitakoa aztertzeko oinarrizko data framea sortzen du
Shiny app batean erabiltzeko egokitua
}
\examples{
# kokapena <- 'extdata/'
# fitxategia <- 'txepetx-2-lo1uvq90o.etherpad'
# o_idat <- f_IDAT(fitxategia)
o_idat

}
