% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_idatzia.R
\name{f_idatzia}
\alias{f_idatzia}
\title{Idazteko orduan sortu den testuaren kalkuluetarako aldagaiak sortzen ditu}
\usage{
f_idatzia(
  objektua,
  lerru = TRUE,
  etenak = TRUE,
  dm = 3000,
  ezabatuak = TRUE,
  copypaste = TRUE,
  lerroko_jauziak = TRUE
)
}
\arguments{
\item{objektua}{IDAT erabilita sortutako zerrenda edo data framea}

\item{lerru}{Aldaketak zein lerrotan ematen diren eta aurreko lerrotik zein distantzia duen}

\item{etenak}{Etenen luzera zein izan den.#'}

\item{dm}{Etentzat hartzeko denbora muga, milisegunduetan. Balio lehenetsia 3 segundu dira, 3000.}

\item{ezabatuak}{Idazketa prozesuan ezabatutakoak zenbat karakterekoak izan diren, pausuka.}

\item{copypaste}{Hamar karaktere baino gehiagoko pausuak nabarmentzen ditu.}

\item{lerroko_jauziak}{Lerroan / paragrafoan bertan egindako lekutze aldaketak karaktereka zenbatzen ditu eta norabidea seinalatzen du.}
}
\value{
Etherpaden idatzitako dokumentu baten \code{objektua}aren prozesuan sortutako
testuaren berreraiketa, hainbat elementu nabarmenduta.
}
\description{
Idazteko orduan sortu den testuaren kalkuluetarako aldagaiak sortzen ditu
}
\examples{
require(dplyr)
require(stringr)
# o_idat <- f_IDAT(fitxategia, kokapena)
f_idatzia(o_idat)$df_aldaketak |>
  filter(egileak == levels(egileak)[2]) |>
  select(idatzia) |>
  unlist() |>
  cat(sep = "")

}
