% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_IDentifikau.R
\name{f_IDentifikau}
\alias{f_IDentifikau}
\title{Idazteko orduan egin diren ekitaldiak identifikatzen ditu}
\usage{
f_IDentifikau(
  objektua,
  egileak = NULL,
  authorka = TRUE,
  nameka = FALSE,
  anonimu = FALSE
)
}
\arguments{
\item{objektua}{IDAT erabilita sortutako zerrenda edo data framea}

\item{egileak}{karaktere bektore bat, authorrek dituen mailen luzerakoa. Egileak
author mailak baino gutxiago badira, izenak errepika daitekez, mailak egokitzeko.}

\item{authorka}{TRUE balioak mailak author horren arabera kalkulatzen ditu.}

\item{nameka}{TRUE balioak mailak name horren arabera kalkulatzen ditu.}

\item{anonimu}{TRUE balioak author edo name-ko balioak E-n zenbakiz ordezkatzen ditu, n zenbaki bat dela.}
}
\value{
IDAT objektu batean egileak batzeko edo anonimatzeko bideak
\code{objektua}aren \code{egileak} eman dakizkio, edo \code{authorka} antolatu,
zein \code{nameka}. Azken aukera bietan anonimo egin daitekez \code{anonimu} aldagaian TRUE emanda.
}
\description{
Idazteko orduan egin diren ekitaldiak identifikatzen ditu
}
\examples{
require(dplyr)
# o_idat <- f_IDAT(kokapena, fitxategia)
f_IDentifikau(o_idat, anonimu = TRUE)$df_aldaketak |>
  select(egileak) |>
  head(10)

}
