% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{o_idat}
\alias{o_idat}
\title{Taldean egindako lan baten analizirako elementuak}
\format{
`o_idat`
Zerrenda bat hiru data frame, testu bat eta autore kode batez:
\describe{
  \item{df_egileak}{Idatzia egin duten idazleak identifikatzeko data framea.}
  \item{df_aldaketak}{Idazketa prozesua}
  \item{df_pool}{edizioetako informazioa}
  \item{pad}{padaren izena}
  \item{ni}{Maiz agertu arren, datu askotan ezabatu beharreko egile bat identifikatzen duena}
  ...
}
}
\source{
<https://www.etherpad.wikimedia.com/p/urlia_artabururen_istorioa >
}
\usage{
o_idat
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
